requirements.txt
nltk
spacy
python-docx
PyPDF2
beautifulsoup4